django/requirements.txt
gunicorn
asgiref==3.4.1
certifi==2023.07.22
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.4
Django==3.2.25
django-appconf==1.0.5
django-bootstrap-form==3.4
django-classy-tags==2.0.0
django-cryptography==1.1
django-filter==21.1
django-formtools==2.4
django-jsonfield==1.4.1
#django-two-factor-auth==1.15.1
django-two-factor-auth[phonenumbers]==1.15.0
#phonenumbers==8.13.5
#phonenumberslite==8.13.5
django-otp==1.2.2
django-phonenumber-field==6.0.0
django-sekizai==2.0.0
django-user-sessions==2.0.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==4.8.3
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.3.7
MarkupSafe==2.0.1
mysqlclient>=2.0,<3
pycparser==2.21
pytz==2022.7.1
qrcode==7.3.1
requests==2.31.0
sentry-sdk==1.43.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.18
zipp==3.6.0
bip329==0.0.3
pymempool==0.0.5
django-datatables-view==1.20.0
django-money==3.3
embit==0.7.0
pycoin==0.92.20230326
